% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockrand2randlist.R
\name{blocked_stratified_randlist}
\alias{blocked_stratified_randlist}
\title{Create randomization lists}
\usage{
blocked_stratified_randlist(
  pi = "",
  acronym = "",
  sample_size = NA,
  seed = NA,
  treatment_levels = c("C", "T"),
  block_size = c(2L, 4L, 6L),
  centres = c("auslre"),
  stratas = centres,
  local_pis = c("PI Cognome Nome (AUSL RE-Irccs)"),
  testing = FALSE,
  print_checks = TRUE,
  export = TRUE,
  export_format = c("pdf", "xlsx", "xlsx_minimal", "envelopes"),
  env_params = list(width = 11, height = 8)
)
}
\arguments{
\item{pi}{global PI}

\item{acronym}{study acronym}

\item{sample_size}{total sample size (a randomization list with
this numerosity will be created for each stratum/element of
stratas)}

\item{seed}{random seed}

\item{treatment_levels}{labels used to identify the groups}

\item{block_size}{blocks dimensions}

\item{centres}{id of involved centres}

\item{stratas}{label for each strata}

\item{local_pis}{a string for footers of the printed lists, in the
same number of stratas}

\item{testing}{if TRUE it will generate a list using a different
seed, for testing purposese (eg EDC setup)}

\item{print_checks}{print performed checks on the randomization
lists}

\item{export}{logical (recicled, by default TRUE), which
randomization lists to export}

\item{export_format}{format of lists exporting (can have pdf, xlsx
and/or envelopes)}

\item{env_params}{envelopes parameters for printing (passed to
blockrand::plotblockrand)}
}
\description{
Create pdf, envelopes and xlsx randomization lists for a
stratified, blocked study and export them
}
