% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockrand2randlist.R
\name{blockrand2randlist}
\alias{blockrand2randlist}
\title{Create a randomization list from a list of blockrand generated data.frame}
\usage{
blockrand2randlist(x, f = "/tmp/randlist", footer = "")
}
\arguments{
\item{x}{named list of blockrand data.frame (names will be used as
sheet name)}

\item{f}{path to file to save in (overwriting the contents). If
NULL the list is displayed.}

\item{footer}{a character vector used as page central
footer(s). Page numbers are on the left, while date/time is on
the right.}
}
\description{
Create a randomization list from a list of blockrand generated data.frame
}
