% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockrand2randlist.R
\name{randlist2envelopes}
\alias{randlist2envelopes}
\title{Create randomization envelopes from a list of blockrand
generated data.frame}
\usage{
randlist2envelopes(
  x = NULL,
  path_prefix = "/tmp/envelopes",
  study_acronym = "",
  env_params = list()
)
}
\arguments{
\item{x}{a single data.frame or a named randlist (that is a
data.frame with id and treatment columns). Names are used for
file naming}

\item{path_prefix}{path prefix of the files to save
in (overwriting the contents).}

\item{study_acronym}{a character vector used as acronym}

\item{env_params}{list of parameters passed to blockrand::plotblockrand}
}
\description{
Create randomization envelopes from a list of blockrand
generated data.frame
}
