% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockrand2randlist.R
\name{randlist2pdf}
\alias{randlist2pdf}
\title{Create pdf randomization lists from a list of blockrand
generated data.frame}
\usage{
randlist2pdf(x, path_prefix = "/tmp/randlist", footer = "")
}
\arguments{
\item{x}{a single data.frame or a named randlist (that is a
data.frame with id and treatment columns). Names are used for
file naming}

\item{path_prefix}{path prefix of the files to save
in (overwriting the contents).}

\item{footer}{a character vector used as page central
footer(s). Must be of length 1 if x is a data.frame or of the
same length of x, if it's a list.}
}
\description{
Create pdf randomization lists from a list of blockrand
generated data.frame
}
